requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.3.0
langchain-huggingface>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
starlette>=0.40.0
httpx>=0.25.0
slowapi>=0.1.9
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.1
transformers>=4.40.0
torch>=2.0.0
scikit-learn>=1.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0
python-json-logger>=2.0.0
prometheus-client>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
redis>=5.0.0